+++ packages/utils/graphql-codegen-testing/src/index.ts
@@ -92,3 +92,4 @@ export function useMonorepo({ dirname }: { dirname: string }) {
 
 export * from './typescript.js';
 export * from './mock-graphql-server.js';
+export * from './resolvers-common.js';