+++ packages/plugins/flow/flow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphql-codegen/flow
 
+## 2.3.6
+
+### Patch Changes
+
+- Updated dependencies [[`63dc8f205`](https://github.com/dotansimha/graphql-code-generator/commit/63dc8f2054e27b944f7d8dc59db8afa85760a127)]:
+  - @graphql-codegen/visitor-plugin-common@2.13.1
+  - @graphql-codegen/plugin-helpers@2.7.2
+
 ## 2.3.5
 
 ### Patch Changes