+++ packages/plugins/java/java/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphql-codegen/java
 
+## 3.3.6
+
+### Patch Changes
+
+- Updated dependencies [[`63dc8f205`](https://github.com/dotansimha/graphql-code-generator/commit/63dc8f2054e27b944f7d8dc59db8afa85760a127)]:
+  - @graphql-codegen/visitor-plugin-common@2.13.1
+  - @graphql-codegen/plugin-helpers@2.7.2
+  - @graphql-codegen/java-common@2.2.6
+
 ## 3.3.5
 
 ### Patch Changes