+++ packages/plugins/typescript/react-query/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @graphql-codegen/typescript-react-query
 
+## 4.0.5
+
+### Patch Changes
+
+- [#8567](https://github.com/dotansimha/graphql-code-generator/pull/8567) [`af40efbd2`](https://github.com/dotansimha/graphql-code-generator/commit/af40efbd23850c6bdb5f5b3d8713bac24e72fe44) Thanks [@EandrewJones](https://github.com/EandrewJones)! - fixes issue #7549
+
+## 4.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`63dc8f205`](https://github.com/dotansimha/graphql-code-generator/commit/63dc8f2054e27b944f7d8dc59db8afa85760a127)]:
+  - @graphql-codegen/visitor-plugin-common@2.13.1
+  - @graphql-codegen/plugin-helpers@2.7.2
+
 ## 4.0.3
 
 ### Patch Changes